refactor(tests): refactoring of all tests
